Enabling dependent sampling
