+++ course/course/course.xml
@@ -21,6 +21,7 @@
         <problem url_name="formula4"/>
         <problem url_name="formula5"/>
         <problem url_name="formula6"/>
+        <problem url_name="formula7"/>
       </vertical>
     </sequential>
     <sequential url_name="singlelistgrader" display_name="SingleListGrader">
