+++ course/problem/formula7.xml
@@ -1,0 +1,24 @@
+<problem display_name="Dependent Sampling" showanswer="always" weight="10" attempts="">
+
+  <p>Often, there are multiple ways of writing a result, depending on the variables you use. Sometimes, you want students to have flexibility in the use of variables. This is particularly common in the case of polar coordinates, where you may want \(x\), \(y\) and \(r\) to appear.</p>
+
+  <p>The answer to the below question is "x/r". You can also enter "x/sqrt(x^2+y^2)".</p>
+
+<script type="text/python" system_path="python_lib">
+from mitxgraders import *
+grader = FormulaGrader(
+    answers='x/r',
+    variables=['x', 'y', 'r'],
+    sample_from={
+        'r': DependentSampler(depends=["x", "y"], formula="sqrt(x^2+y^2)")
+    }
+)
+</script>
+
+<customresponse cfn="grader" inline="1">
+  <textline math="1" inline="1" correct_answer="x/r"/>
+</customresponse>
+
+<a href="https://github.com/mitodl/mitx-grading-library/tree/master/course/problem/formula7.xml" target="_blank">View source</a>
+
+</problem>
